+++ src/Http/Requests/FastlaneRequest.php
@@ -2,9 +2,14 @@
 
 namespace CbtechLtd\Fastlane\Http\Requests;
 
+use CbtechLtd\Fastlane\QueryFilter\QueryFilter;
+use CbtechLtd\Fastlane\QueryFilter\QueryFilterContract;
 use CbtechLtd\Fastlane\Support\Contracts\EntryInstance as EntryInstanceContract;
 use CbtechLtd\Fastlane\Support\Contracts\EntryType as EntryTypeContract;
+use CbtechLtd\Fastlane\Support\Contracts\SchemaField;
 use Illuminate\Http\Request;
+use Illuminate\Support\Collection;
+use Illuminate\Support\Str;
 
 class FastlaneRequest extends Request
 {
@@ -38,5 +43,23 @@ class FastlaneRequest extends Request
     public function getEntryInstance(): EntryInstanceContract
     {
         return $this->entryInstance;
+    }
+
+    public function buildQueryFilter(): QueryFilterContract
+    {
+        return tap(new QueryFilter(), function (QueryFilter $qf) {
+
+            if ($order = $this->input('order')) {
+                $fieldName = Str::replaceFirst('-', '', $order);
+
+                $sortable = Collection::make($this->getEntryInstance()->schema()->getIndexFields())
+                    ->filter(fn(SchemaField $field) => $field->isSortable())
+                    ->first(fn(SchemaField $field) => $field->isSortable() && $field->getName() === $fieldName);
+
+                if ($sortable) {
+                    $qf->withOrder($order);
+                }
+            }
+        });
     }
 }